cmd/tractor: extract command code as agentsystrayservice package
